+++ release-notes.md
@@ -1,5 +1,10 @@
 # arthurpicht/utils-core release notes
 
+## v1.15.0 2024-05-17
+
+* ISODates.toISO(long epochMillis)
+* ISODates.toLocalISO(long epochMillis)
+
 ## v1.14.0 2024-05-17
 
 * strings.Timestamps moved to dates.TimestampStrings